refactor(storage): move eviction logic into its own module
Pull SweepStats / EvictStats / StartupCleanupStats and the sweep_orphans
/ evict_to_size / evict_all / run_startup_cleanup family out of
storage/service.rs into a new storage/eviction.rs module.

StorageService keeps a thin delegate API so external callers (lib.rs,
commands::clear_cache) compile unchanged; the previously-private
audio_dir and entries fields are pub(super) so the sibling module can
reach them, and the cleanup-related tests move with their code.

Also fix a clippy::zero_prefixed_literal in the eviction tests
(0_001 → 1) so cargo clippy --all-targets stays clean.
