+++ src-tauri/src/storage/mod.rs
@@ -1,2 +1,3 @@
+pub mod eviction;
 pub mod schema;
 pub mod service;
